Fix build issues on UCRT64/MinGW and MSVC environments
- Comment out `LIBRARY` in `migemo.def` so that the generated DLL name
  and import library reference match across different toolchains (e.g.,
  `libmigemo.dll` on UCRT64 vs `migemo.dll` on MSVC).
- Define `S_ISREG` macro if missing in MSVC environments.
- Add explicit `(int)` casts in `wordbuf.c` to prevent implicit truncation
  warnings from `size_t` to `int` on 64-bit builds.

+++ src/wordbuf.c
@@ -133,11 +133,11 @@ wordbuf_write_bytes(wordbuf_p buf, const unsigned char *p, size_t len)
 {
     if (p != NULL && len > 0)
     {
-        int newlen = buf->last + len + 1;
+        int newlen = buf->last + (int)len + 1;
         if (newlen > buf->len && !wordbuf_extend(buf, newlen))
             return 0;
         memcpy(&buf->buf[buf->last], p, len);
-        buf->last = buf->last + len;
+        buf->last = buf->last + (int)len;
         buf->buf[buf->last] = '\0';
     }
     return buf->last;
